+++ src/Tabs/index.js
@@ -1,0 +1,5 @@
+import Tab from './Tab'
+import Tabs from './Tabs'
+Tabs.Tab = Tab
+
+export default Tabs