+++ CHANGELOG.md
@@ -2,6 +2,36 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.0.19](https://github.com/voxelcloud/vc-ui/compare/v0.0.17...v0.0.19) (2021-04-11)
+
+
+### Features
+
+* formUpdater component ([c6e8a64](https://github.com/voxelcloud/vc-ui/commit/c6e8a64c621dbd76480129438201719bce456884))
+* formValidator component ([b40d30d](https://github.com/voxelcloud/vc-ui/commit/b40d30d79a3fbf83bda2945886eadbf6398e0094))
+* textFiled component add readOnly props ([45de232](https://github.com/voxelcloud/vc-ui/commit/45de23236a18ce8123f6fdf1f10d5c15461cbd0c))
+
+
+### Bug Fixes
+
+* tabs component fix width style ([d4ce7c5](https://github.com/voxelcloud/vc-ui/commit/d4ce7c5d8b4356aa6be55a9ffc3ed65454871273))
+* textField label color not change when focused in readOnly mode ([73df8cb](https://github.com/voxelcloud/vc-ui/commit/73df8cb92a67d263eeaa229fcfaa214f6fc8348d))
+
+### [0.0.18](https://github.com/voxelcloud/vc-ui/compare/v0.0.16...v0.0.18) (2021-04-01)
+
+
+### Features
+
+* textFiled component add readOnly props ([45de232](https://github.com/voxelcloud/vc-ui/commit/45de23236a18ce8123f6fdf1f10d5c15461cbd0c))
+* util ([5b2e537](https://github.com/voxelcloud/vc-ui/commit/5b2e537f8556f5573450c3243066c6a50dee7984))
+* util ([a6e904b](https://github.com/voxelcloud/vc-ui/commit/a6e904b26dcfeaac07b39756cab468f0e7433ae5))
+* **util:** add common util ([9b4255d](https://github.com/voxelcloud/vc-ui/commit/9b4255d23909a97700415500ccd25a888a60d084))
+
+
+### Bug Fixes
+
+* voxios error handler ([3345f31](https://github.com/voxelcloud/vc-ui/commit/3345f3186c7e3ea2432ae69773237c3e6a7c0c48))
+
 ### [0.0.17](https://github.com/voxelcloud/vc-ui/compare/v0.0.7...v0.0.17) (2021-03-19)